refactor: rename CreationForm as it is used for both creation and edition
